+++ cmd/main.go
@@ -4,12 +4,12 @@ import (
 	"fmt"
 	"os"
 
+	"github.com/maksimboikodev/test/cmd/pkg/usecases"
+
 	"github.com/kelseyhightower/envconfig"
 	"github.com/sirupsen/logrus"
 )
 
-type Celsius float32
-type Fahrenheit float32
 type EnvConfig struct {
 	LogFile string `envconfig:"LOG_FILE"`
 }
@@ -17,6 +17,7 @@ type EnvConfig struct {
 func main() {
 	var eConf EnvConfig
 	envconfig.Process("", &eConf)
+
 	file, _ := os.OpenFile(eConf.LogFile, os.O_CREATE|os.O_APPEND|os.O_WRONLY, 0666)
 
 	var log = logrus.New()
@@ -27,13 +28,10 @@ func main() {
 	fmt.Scan(&d)
 	fmt.Println(d)
 
-	c := Celsius(d)
-	log.Info("температура по фаренгейту  ", toFahrenheit(c))
+	tempHistoryHandler := &usecases.HistoryHander{
+		Logger:  log,
+		Samples: []usecases.Fahrenheit{},
+	}
 
-}
-
-func toFahrenheit(t Celsius) Fahrenheit {
-	var temp Fahrenheit
-	temp = Fahrenheit((t * 9 / 5) + 32)
-	return temp
+	usecases.CheckandSave(tempHistoryHandler, usecases.Celsius(d), 5)
 }
